In process of implementing create_all_for_number_of_symbols function.

+++ src/symbols/Function.rs
@@ -5,10 +5,10 @@ pub mod Subtraction;
 pub mod SquareRoot;
 
 pub trait Function: Symbol {
-    fn min_variables(&self) -> i16 {
+    fn min_variables(&self) -> i128 {
         1
     }
-    fn max_variables(&self) -> i16 {
+    fn max_variables(&self) -> i128 {
         0 // 0 means the function has no limit on the number of variable it can process
     }
     fn compute(&self, variables: Vec<i128>) -> i128;
