+++ src/symbols/Function/Addition.rs
@@ -10,7 +10,7 @@ impl Symbol for Addition {
 }
 
 impl Function for Addition {
-    fn min_variables(&self) -> i16 {
+    fn min_variables(&self) -> i128 {
         return 2;
     }
     fn compute(&self, variables: Vec<i128>) -> i128 {
